requirements.txt
appdirs==1.4.3
bleach==2.0.0
decorator==4.0.11
defusedxml==0.5.0
Django==1.9.12
django-autocomplete-light==2.3.3
django-crispy-forms==1.6.1
django-extensions==1.7.7
django-filter==1.0.1
django-gm2m==0.4.8
django-guardian==1.4.6
django-registration==2.2
django-reversion==2.0.8
django-reversion-compare==0.7.3
django-tables2==1.4.2
djangorestframework==3.6.2
djangorestframework-csv==2.0.0
djangorestframework-xml==1.3.0
entrypoints==0.2.2
html5lib==0.999999999
ipykernel==4.6.0
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
isodate==0.5.4
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.0
jupyter-console==5.1.0
jupyter-core==4.3.0
MarkupSafe==1.0
mistune==0.7.4
mysqlclient==1.3.10
nbconvert==5.1.1
nbformat==4.3.0
nltk==3.2.2
notebook==4.4.1
numpy==1.12.1
packaging==16.8
pandas==0.19.2
pandocfilters==1.4.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.0
rdflib==4.2.2
requests==2.13.0
scikit-learn==0.18.1
scipy==0.19.0
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
terminado==0.6
testpath==0.3
tornado==4.4.3
traitlets==4.3.2
unicodecsv==0.14.1
wcwidth==0.1.7
webencodings==0.5
widgetsnbextension==2.0.0
